Add our Python virtual environment runner script
This is a bash script designed to bridge between the regular shell
environment and venv environment. Copy this script to any venv
directory, and symlink to it using the name of the desired command to be
run from anywhere on your path. Then simply running the command will
first activate the venv, and then run the command in that venv.

See runner's comments for more detail and an example.