Add AsciiString class to handy.py
An AsciiString is a regular string that translates its characers from
Unicode to ASCII if possible. Any untranslatable characters are left in
the string unchnaged, which you can test for using the isascii() method.

I've also finally removed the decline_deprecated() function from this
module. See the english.py module in this repo for its far-superior
replacement.

I've escaped the \r characters in class Spinner's docstring so they
render as "\r" in pydoc and polished this module's help text a bit.