More work on installing virtual environments
Minor bug fix. The symlinks to "base" weren't being created in the right
place. I'm actually still working on this one, but it's late and I want
to check the code in.

And I've added a debug_func function decorator in debug.py. I still need
to parameterize it, but it's useful enough in a quick-and-dirty way
right now. The rest of the changes were just cosmetic.