Optimize AsciiString's translator for editability
Previous versions mapped each translatable non-ASCII character to the
corresponding ASCII character. Because many non-ASCII characters might
map to one ASCII character, editing the map was more complicated than
necessary.

This new version maps each ASCII target character to a set of the
non-ASCII characters that translate to it. The algorighm for creating
string values for str.maketrans() is a little more involved, but this
happens only once during the life of a given process, and the map itself
is much more human-readable and -editable.