Add --started option to the progress script
Use an argument of the form "[[YY]YY-MM-DD ] [HH:MM[:SS]" to tell
--started when the process being monitored was started. The presence of
this option turns on the ETA at the end of the output value.

The ETA is expressed as "DdHhMmSs" using a maximum of two contiguous
components of that. So you might get something like "1d8h", "8h47m",
"27m55s", or just "32s". Computing this ETA is entirely linear. If the
time spent on each unit of work varies in a way that diverges from the
mean, the ETA will be unreliable.