Lots of changes the "progress" script
Read "progress -h" for the details, but I've greatly improved the
internals and have tweaked the command line options a bit. I think the
script itself is ready for prime time, and I'll likely be separating
this into a main script and a progress.py module before long so I can
use this from Python as well as from the shell.