Add pyparsing back into our installer
We're now installing pyparsing using pip (like I suppose I should have
done in the first place).

I've also updated Command.__call__() in pylib/install.py to not raise an
exception if the command exits with a code other than 0. The result is
avilable from the (possible failed) Command instance if the caller wants
to checkk it.