Intermediate enhancement of the install script
I'm adding Python virtual environments to the things this script knows
how to install, but it's not working yet. (Look for the RESUME point.)

I've made too many changes to lose, so I'm committing this intermediate
version. If you have to install something before this in finished,
you'll have to comment out the stuff that crashes.