Update pwgen's "pronounceable" algorithm
The vowels e and o now have a 1-in-20 chance of being doubled in a
sylable. The consonants b, d, f, g, l, m, n, p, s, t have a 1-in-30
chance of being doubled at the start of a second-or-subsequent sylable.
This gives pronounceable passwords more variety without diminishing
their pronounceability.